[EuiPortal] Add global settings for Portal-Based Components
