+++ src/components/portal/portal.types.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import { keysOf } from '../common';
+
+interface InsertPositionsMap {
+  after: InsertPosition;
+  before: InsertPosition;
+}
+
+export const insertPositions: InsertPositionsMap = {
+  after: 'afterend',
+  before: 'beforebegin',
+};
+
+export type EuiPortalInsertPosition = keyof typeof insertPositions;
+
+export const INSERT_POSITIONS: EuiPortalInsertPosition[] =
+  keysOf(insertPositions);
+
+export interface EuiPortalInsertion {
+  sibling: HTMLElement;
+  position: EuiPortalInsertPosition;
+}
